requirements.txt
# Core AI/ML libraries
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
google-generativeai>=0.3.0
chromadb>=0.4.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# PDF + LLM helpers
PyMuPDF>=1.23.0
python-docx>=1.0.1
instructor>=0.5.4
tiktoken>=0.5.1

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Web framework (if switching from Streamlit)
flask>=3.0.0
flask-cors>=4.0.0

# Visualization (if used later)
networkx>=3.0
matplotlib>=3.7.0
plotly>=5.15.0

# Streamlit frontend
streamlit>=1.32.0

# Export functionality
openpyxl>=3.1.0
reportlab>=4.0.0
PyPDF2>=3.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Dev/testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

langchain-community>=0.0.21
